Fix version numbers. Rearranged the inventory transactions a bit.
