+++ src/MiNET/MiNET/LoginMessageHandler.cs
@@ -371,7 +371,7 @@ namespace MiNET
 			IMcpeMessageHandler messageHandler = server.CreatePlayer(_session, _playerInfo);
 			_session.MessageHandler = messageHandler; // Replace current message handler with real one.
 
-			if (_playerInfo.ProtocolVersion < 111)
+			if (_playerInfo.ProtocolVersion < 137 || _playerInfo.ProtocolVersion > 65535)
 			{
 				Log.Warn($"Wrong version ({_playerInfo.ProtocolVersion}) of Minecraft. Upgrade to join this server.");
 				_session.Disconnect($"Wrong version ({_playerInfo.ProtocolVersion}) of Minecraft. Upgrade to join this server.");
