+++ src/MiNET/MiNET/MotdProvider.cs
@@ -25,7 +25,7 @@ namespace MiNET
 			NumberOfPlayers = serverInfo.NumberOfPlayers;
 			MaxNumberOfPlayers = serverInfo.MaxNumberOfPlayers;
 
-			var protocolVersion = "120";
+			var protocolVersion = "137";
 			var clientVersion = "1.2.0";
 			var edition = "MCPE";
