+++ src/MiNET/MiNET/Utils/Transaction.cs
@@ -1,0 +1,84 @@
+#region LICENSE
+
+// The contents of this file are subject to the Common Public Attribution
+// License Version 1.0. (the "License"); you may not use this file except in
+// compliance with the License. You may obtain a copy of the License at
+// https://github.com/NiclasOlofsson/MiNET/blob/master/LICENSE. 
+// The License is based on the Mozilla Public License Version 1.1, but Sections 14 
+// and 15 have been added to cover use of software over a computer network and 
+// provide for limited attribution for the Original Developer. In addition, Exhibit A has 
+// been modified to be consistent with Exhibit B.
+// 
+// Software distributed under the License is distributed on an "AS IS" basis,
+// WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License for
+// the specific language governing rights and limitations under the License.
+// 
+// The Original Code is Niclas Olofsson.
+// 
+// The Original Developer is the Initial Developer.  The Initial Developer of
+// the Original Code is Niclas Olofsson.
+// 
+// All portions of the code written by Niclas Olofsson are Copyright (c) 2014-2017 Niclas Olofsson. 
+// All Rights Reserved.
+
+#endregion
+
+using System.Collections.Generic;
+using System.Numerics;
+using MiNET.Items;
+using MiNET.Net;
+
+namespace MiNET.Utils
+{
+	public class Transaction
+	{
+		public McpeInventoryTransaction.TransactionType TransactionType { get; set; }
+
+		public List<TransactionRecord> Transactions { get; set; } = new List<TransactionRecord>();
+
+		public int ActionType { get; set; }
+		public BlockCoordinates Position { get; set; }
+		public int Face { get; set; }
+		public int Slot { get; set; }
+		public Item Item { get; set; }
+		public Vector3 FromPosition { get; set; }
+		public Vector3 ClickPosition { get; set; }
+		public long EntityId { get; set; }
+
+		public Transaction()
+		{
+		}
+	}
+
+	public abstract class TransactionRecord
+	{
+		public int Source { get; set; }
+		public int Slot { get; set; }
+		public Item OldItem { get; set; }
+		public Item NewItem { get; set; }
+	}
+
+	public class ContainerTransactionRecord : TransactionRecord
+	{
+		public int InventoryId { get; set; }
+	}
+
+	public class GlobalTransactionRecord : TransactionRecord
+	{
+	}
+
+	public class WorldInteractionTransactionRecord : TransactionRecord
+	{
+		public int Flags { get; set; } // NoFlag = 0 WorldInteractionRandom = 1
+	}
+
+	public class CreativeTransactionRecord : TransactionRecord
+	{
+		public int InventoryId { get; set; } = 0x79; // Creative
+	}
+
+	public class CraftTransactionRecord : TransactionRecord
+	{
+		public int Action { get; set; }
+	}
+}
